refactor(editor): move all plugin menu utils to rows plugin
